Upgrade Mirai to 2.5.0

+++ src/main/java/com/github/cookiesjuice/cookiesbot/control/SetuListenControl.kt
@@ -10,9 +10,13 @@ import com.github.cookiesjuice.cookiesbot.module.setu.service.UserService
 import com.github.cookiesjuice.cookiesbot.module.setu.service.impl.ImageServiceImpl
 import com.github.cookiesjuice.cookiesbot.utils.HttpUtils
 import net.mamoe.mirai.Bot
-import net.mamoe.mirai.event.subscribeGroupMessages
-import net.mamoe.mirai.event.subscribeMessages
+import net.mamoe.mirai.event.GlobalEventChannel
+import net.mamoe.mirai.event.events.GroupMessageEvent
+import net.mamoe.mirai.event.events.MessageEvent
 import net.mamoe.mirai.message.data.*
+import net.mamoe.mirai.message.data.Image.Key.queryUrl
+import net.mamoe.mirai.message.data.MessageSource.Key.quote
+import net.mamoe.mirai.utils.ExternalResource.Companion.toExternalResource
 import org.springframework.stereotype.Controller
 import java.io.File
 import java.util.*
@@ -20,21 +24,21 @@ import java.util.regex.Pattern
 
 @Controller
 class SetuListenControl(
-        val cmdService: CmdService,
-        val setuService: SetuService,
-        val userService: UserService,
-        val imageService: ImageServiceImpl,
-        val botProperties: BotProperties,
-        val setuProperties: SetuProperties,
-        val userProperties: UserProperties,
+    val cmdService: CmdService,
+    val setuService: SetuService,
+    val userService: UserService,
+    val imageService: ImageServiceImpl,
+    val botProperties: BotProperties,
+    val setuProperties: SetuProperties,
+    val userProperties: UserProperties,
 ) {
     /**
      * 上传涩图 | 获取一张指定id的涩图 | 随机获取一张或多张涩图
      */
-    private fun Bot.setu(): Boolean {
+    private fun setu(): Boolean {
         var isHandle = false
-        subscribeGroupMessages {
-            contains("涩图") {
+        GlobalEventChannel.subscribeAlways<GroupMessageEvent> {
+            if (message.content.contains("涩图")) {
                 val matcher = Pattern.compile("(上传)*(涩图[\\s]*)([\\d+]*)(\\s*连)*").matcher(message.content)
                 if (matcher.find()) {
                     val user = userService.findOrSave(sender.id)
@@ -53,7 +57,7 @@ class SetuListenControl(
                     if (hasG1) {//上传涩图
                         val image = message[Image]
                         if (image != null) {
-                            val url = image.url()
+                            val url = image.queryUrl()
                             val path = File("temp")
                             if (!path.exists()) {
                                 path.mkdir()
@@ -61,7 +65,7 @@ class SetuListenControl(
                             val saveName = "$path/${UUID.randomUUID()}"
                             HttpUtils.downloadFile(url, saveName)
                             val tempFile = File(saveName)
-                            quoteReply("正在上传中喵~")
+                            message.quote().plus("正在上传中喵~").sendTo(subject)
                             try {
                                 val setu = setuService.upload(user, tempFile)
                                 if (setu != null) {
@@ -69,28 +73,29 @@ class SetuListenControl(
 
                                     val file = imageService.evaluates(setu.evaluates)
                                     if (file != null) {
-                                        quoteReply(msg.plus("分析标签：\n").plus(uploadImage(file)))
+                                        val img = subject.uploadImage(file.toExternalResource())
+                                        message.quote().plus(msg.plus("分析标签：\n").plus(img)).sendTo(subject)
                                         file.delete()
                                     } else {
-                                        quoteReply(msg.plus("哎呀~分析结果被曲奇吃掉惹喵~"))
+                                        message.quote().plus(msg.plus("哎呀~分析结果被曲奇吃掉惹喵~")).sendTo(subject)
                                     }
 
                                     userService.changeExp(user, setuProperties.uploadExp.toLong())
                                 } else {
-                                    quoteReply("上传失败惹~涩图被曲奇吃掉惹喵~")
+                                    message.quote().plus("上传失败惹~涩图被曲奇吃掉惹喵~").sendTo(subject)
                                 }
                             } finally {
                                 tempFile.delete()
                             }
                         } else {
-                            quoteReply("上传个喵喵?没有看到涩图喵~")
+                            message.quote().plus("上传个喵喵?没有看到涩图喵~").sendTo(subject)
                         }
 
                     } else if (hasG3 && hasG4 && !isMaximum) { //涩图xxx连
                         try {
                             val num = g3.toInt()
                             if (num > setuProperties.setuMax) {
-                                quoteReply("最多只能连续${setuProperties.setuMax}张涩图喵~")
+                                message.quote().plus("最多只能连续${setuProperties.setuMax}张涩图喵~").sendTo(subject)
                             } else if (num + everyday.setuCount > max) {
                                 isMaximum = true
                             } else {
@@ -101,19 +106,19 @@ class SetuListenControl(
                                         if (setu != null) {
                                             val setuFile = setuService.getFile(setu)
                                             if (setuFile != null) {
-                                                val image: Image = uploadImage(setuFile)
+                                                val image: Image = subject.uploadImage(setuFile.toExternalResource())
                                                 reply = reply.plus(image)
                                             }
                                         }
                                     }
                                     userService.addSetuCount(everyday, num)
-                                    quoteReply(reply)
+                                    message.quote().plus(reply).sendTo(subject)
                                 } else {
                                     throw NumberFormatException()
                                 }
                             }
                         } catch (_: NumberFormatException) {
-                            quoteReply("没有找到涩图喵~")
+                            message.quote().plus("没有找到涩图喵~").sendTo(subject)
                         }
                     } else if (!isMaximum) { //涩图
                         //如果没超过此等级每日获取次数
@@ -132,22 +137,22 @@ class SetuListenControl(
                         if (setu != null) {
                             val setuFile = setuService.getFile(setu)
                             if (setuFile != null) {
-                                val image: Image = uploadImage(setuFile)
+                                val image: Image = subject.uploadImage(setuFile.toExternalResource())
                                 val info = "id: [${setu.id}]"
-                                quoteReply(messageChainOf(PlainText(info)).plus(image))
+                                message.quote().plus(messageChainOf(PlainText(info)).plus(image)).sendTo(subject)
                                 userService.addSetuCount(everyday)
                                 hasSetu = true
                             }
                         }
 
                         if (!hasSetu) {
-                            quoteReply("没有找到涩图喵~")
+                            message.quote().plus("没有找到涩图喵~").sendTo(subject)
                         }
                     }
 
                     if (isMaximum) {
                         val tips = "你的等级为[" + user.level + "]级，每天最多只能获取[" + max + "]张涩图，今日已经获取[" + everyday.setuCount + "]张涩图了哦喵~"
-                        quoteReply(tips)
+                        message.quote().plus(tips).sendTo(subject)
                     }
                     isHandle = true
                 }
@@ -160,15 +165,16 @@ class SetuListenControl(
     /**
      * 对涩图评分，或收藏涩图
      */
-    private fun Bot.reply(): Boolean {
+    private fun reply(): Boolean {
         var isHandle = false
-        subscribeGroupMessages {
+        GlobalEventChannel.subscribeAlways<GroupMessageEvent> {
             //如果是引用回复，检查对涩图评分，收藏事件
-            has<QuoteReply> {
-                if (it.source.fromId == id) {
+            val quote = message[QuoteReply]
+            if (quote != null) {
+                if (quote.source.fromId == bot.id) {
                     val user = userService.findOrSave(sender.id)
                     val pattern = Pattern.compile("(id: \\[)(\\d+)(])")
-                    val matcher = pattern.matcher(it.source.originalMessage[PlainText]!!.contentToString())
+                    val matcher = pattern.matcher(quote.source.originalMessage.contentToString())
                     if (matcher.find()) {
                         val id = matcher.group(2).toLong()
                         val setu = setuService.find(id)
@@ -176,9 +182,9 @@ class SetuListenControl(
                             val centent = message.contentToString()
                             if (centent.contains("喜欢") || centent.contains("收藏")) {
                                 if (userService.addFavorite(user, setu)) {
-                                    quoteReply("收藏涩图[${setu.id}]成功喵~")
+                                    message.quote().plus("收藏涩图[${setu.id}]成功喵~").sendTo(subject)
                                 } else {
-                                    quoteReply("已经收藏过此涩图了喵~")
+                                    message.quote().plus("已经收藏过此涩图了喵~").sendTo(subject)
                                 }
                             } else {
                                 val err = "喵?不明白你的意思喵~"
@@ -197,21 +203,21 @@ class SetuListenControl(
                                             val oldScore = setuService.grade(user, setu, score)
                                             if (oldScore != 0) {
                                                 if (oldScore == score) {
-                                                    quoteReply("未更新评分，本次评分与上次评分是一样的喵~")
+                                                    message.quote().plus("未更新评分，本次评分与上次评分是一样的喵~").sendTo(subject)
                                                 } else {
-                                                    quoteReply("评分已更新喵~ $oldScore -> $score")
+                                                    message.quote().plus("评分已更新喵~ $oldScore -> $score").sendTo(subject)
                                                 }
                                             } else {
-                                                quoteReply("评分成功喵~")
+                                                message.quote().plus("评分成功喵~").sendTo(subject)
                                             }
                                         } else {
-                                            quoteReply("分数超出范围(1~100)了喵~")
+                                            message.quote().plus("分数超出范围(1~100)了喵~").sendTo(subject)
                                         }
                                     } else {
-                                        quoteReply(err)
+                                        message.quote().plus(err).sendTo(subject)
                                     }
                                 } catch (e: NumberFormatException) {
-                                    quoteReply(err)
+                                    message.quote().plus(err).sendTo(subject)
                                 }
                             }
                         }
@@ -226,29 +232,29 @@ class SetuListenControl(
     /**
      * 分析图片标签
      */
-    private fun Bot.evaluate(): Boolean {
+    private fun evaluate(): Boolean {
         var isHandle = false
-        subscribeGroupMessages {
-            has<Image> {
+        GlobalEventChannel.subscribeAlways<GroupMessageEvent> {
+            val image = message[Image]
+            if (image != null) {
                 if (message.contentToString().contains("分析")) {
                     val path = File("temp")
                     if (!path.exists()) {
                         path.mkdir()
                     }
                     val saveName = "$path/${UUID.randomUUID()}"
-                    HttpUtils.downloadFile(it.url(), saveName)
+                    HttpUtils.downloadFile(image.queryUrl(), saveName)
                     val tempFile = File(saveName)
-                    quoteReply("正在分析中喵~")
+                    message.quote().plus("正在分析中喵~").sendTo(subject)
                     val evaluateList = setuService.evaluate(tempFile.path)
                     tempFile.delete()
-                    val info = PlainText("分析完成喵~\n")
-
                     val file = imageService.evaluates(evaluateList)
                     if (file != null) {
-                        quoteReply(info.plus(uploadImage(file)))
+                        val img = subject.uploadImage(file.toExternalResource())
+                        message.quote().plus("分析完成喵~\n").plus(img).sendTo(subject)
                         file.delete()
                     } else {
-                        quoteReply("哎呀~分析结果被曲奇吃掉惹喵~")
+                        message.quote().plus("哎呀~分析结果被曲奇吃掉惹喵~").sendTo(subject)
                     }
 
                     isHandle = true
@@ -261,98 +267,95 @@ class SetuListenControl(
     /**
      * 发言次数记录并增加经验
      */
-    private fun Bot.speak() {
-        subscribeGroupMessages {
-            always {
-                val user = userService.findOrSave(sender.id)
-                val everyday = userService.getToday(user)
+    private fun speak() {
+        GlobalEventChannel.subscribeAlways<GroupMessageEvent> {
+            val user = userService.findOrSave(sender.id)
+            val everyday = userService.getToday(user)
 
-                //每次发言都会增加经验
-                //如果发言加经验的次数没超过配置的最大值
-                if (everyday.speakCount < userProperties.speakExpMaxOfDay) {
-                    userService.changeExp(user, userProperties.speakExp.toLong())
-                }
-
-                //增加发言计数
-                userService.addSpeakCount(everyday)
+            //每次发言都会增加经验
+            //如果发言加经验的次数没超过配置的最大值
+            if (everyday.speakCount < userProperties.speakExpMaxOfDay) {
+                userService.changeExp(user, userProperties.speakExp.toLong())
             }
+
+            //增加发言计数
+            userService.addSpeakCount(everyday)
         }
     }
 
-    private fun Bot.info(): Boolean {
+    private fun info(): Boolean {
         var isHandle = false
-        subscribeMessages {
-            always {
-                val content = message.contentToString()
-                if (content.contains("myinfo")) {
-                    val user = userService.findOrSave(sender.id)
-                    val file = imageService.user(user)
-                    if (file != null) {
-                        quoteReply(uploadImage(file))
-                        file.delete()
+        GlobalEventChannel.subscribeAlways<MessageEvent> {
+            val content = message.contentToString()
+            if (content.contains("myinfo")) {
+                val user = userService.findOrSave(sender.id)
+                val file = imageService.user(user)
+                if (file != null) {
+                    val img = subject.uploadImage(file.toExternalResource())
+                    message.quote().plus(img).sendTo(subject)
+                    file.delete()
+                } else {
+                    message.quote().plus("哎呀~消息被曲奇吃掉惹喵~").sendTo(subject)
+                }
+                isHandle = true
+            } else {
+                val pattern = Pattern.compile("(setuinfo\\s*)(\\d+)")
+                val matcher = pattern.matcher(message.contentToString())
+                if (matcher.find()) {
+                    val setu = try {
+                        val id = matcher.group(2).toLong()
+                        setuService.find(id)
+                    } catch (_: NumberFormatException) {
+                        null
+                    }
+                    if (setu != null) {
+                        val file = imageService.setu(setu)
+                        if (file != null) {
+                            val img = subject.uploadImage(file.toExternalResource())
+                            message.quote().plus(img).sendTo(subject)
+                            file.delete()
+                        } else {
+                            message.quote().plus("哎呀~消息被曲奇吃掉惹喵~").sendTo(subject)
+                        }
                     } else {
-                        quoteReply("哎呀~消息被曲奇吃掉惹喵~")
+                        message.quote().plus("没有找到涩图喵~").sendTo(subject)
                     }
                     isHandle = true
-                } else {
-                    val pattern = Pattern.compile("(setuinfo\\s*)(\\d+)")
-                    val matcher = pattern.matcher(message.contentToString())
-                    if (matcher.find()) {
-                        val setu = try {
-                            val id = matcher.group(2).toLong()
-                            setuService.find(id)
-                        } catch (_: NumberFormatException) {
-                            null
-                        }
-                        if (setu != null) {
-                            val file = imageService.setu(setu)
-                            if (file != null) {
-                                quoteReply(uploadImage(file))
-                                file.delete()
-                            } else {
-                                quoteReply("哎呀~消息被曲奇吃掉惹喵~")
-                            }
-                        } else {
-                            quoteReply("没有找到涩图喵~")
-                        }
-                        isHandle = true
-                    }
                 }
             }
         }
+
         return isHandle
     }
 
-    private fun Bot.cmd(): Boolean {
+    private fun cmd(): Boolean {
         var isHandle = false
-        subscribeMessages {
-            always {
-                val pattern = Pattern.compile("^(${botProperties.cmd})(\\S+)([\\s+\\S]*)")
-                val matcher = pattern.matcher(message.content)
-                if (matcher.find()) {
-                    val cmd = matcher.group(2)
-                    val args = arrayListOf<String>()
-                    matcher.group(3).split("\\s".toRegex()).forEach { s ->
-                        if (s != "") {
-                            args.add(s)
-                        }
+        GlobalEventChannel.subscribeAlways<MessageEvent> {
+            val pattern = Pattern.compile("^(${botProperties.cmd})(\\S+)([\\s+\\S]*)")
+            val matcher = pattern.matcher(message.content)
+            if (matcher.find()) {
+                val cmd = matcher.group(2)
+                val args = arrayListOf<String>()
+                matcher.group(3).split("\\s".toRegex()).forEach { s ->
+                    if (s != "") {
+                        args.add(s)
                     }
-                    val reply = cmdService.handleCmd(cmd, args.toTypedArray(), sender.id)
-                    quoteReply(reply)
-                    isHandle = true
                 }
+                val reply = cmdService.handleCmd(cmd, args.toTypedArray(), sender.id)
+                message.quote().plus(reply).sendTo(subject)
+                isHandle = true
             }
         }
         return isHandle
     }
 
     fun listen(bot: Bot) {
-        bot.speak()
+        speak()
 
-        if (bot.cmd()) return
-        if (bot.info()) return
-        if (bot.reply()) return
-        if (bot.setu()) return
-        if (bot.evaluate()) return
+        if (cmd()) return
+        if (info()) return
+        if (reply()) return
+        if (setu()) return
+        if (evaluate()) return
     }
 }
